requirements.txt
langchain>=0.0.350
langchain-core>=0.1.16
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.6.1
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.9.0
pydantic>=2.5.3


